+++ pymeasure/instruments/keithley/keithley6221.py
@@ -90,7 +90,7 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     source_enabled = Instrument.control(
         "OUTPut?", "OUTPut %d",
-        """A boolean property that controls whether the source is enabled, takes
+        """Control (boolean) whether the source is enabled, takes
         values True or False. The convenience methods :meth:`~.Keithley6221.enable_source` and
         :meth:`~.Keithley6221.disable_source` can also be used.""",
         validator=strict_discrete_set,
@@ -109,7 +109,7 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     source_delay = Instrument.control(
         ":SOUR:DEL?", ":SOUR:DEL %g",
-        """ A floating point property that sets a manual delay for the source
+        """ Control (floating) a manual delay for the source
         after the output is turned on before a measurement is taken. When this
         property is set, the auto delay is turned off. Valid values are
         between 1e-3 [seconds] and 999999.999 [seconds].""",
@@ -119,7 +119,7 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     output_low_grounded = Instrument.control(
         ":OUTP:LTE?", "OUTP:LTE %d",
-        """ A boolean property that controls whether the low output of the triax
+        """ Control (boolean) whether the low output of the triax
         connection is connected to earth ground (True) or is floating (False). """,
         validator=strict_discrete_set,
         values={True: 1, False: 0},
@@ -132,20 +132,20 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     source_current = Instrument.control(
         ":SOUR:CURR?", ":SOUR:CURR %g",
-        """ A floating point property that controls the source current
+        """ Control (floating) the source current
         in Amps. """,
         validator=truncated_range,
         values=[-0.105, 0.105]
     )
     source_compliance = Instrument.control(
         ":SOUR:CURR:COMP?", ":SOUR:CURR:COMP %g",
-        """A floating point property that controls the compliance of the current
+        """Control (floating) the compliance of the current
         source in Volts. valid values are in range 0.1 [V] to 105 [V].""",
         validator=truncated_range,
         values=[0.1, 105])
     source_range = Instrument.control(
         ":SOUR:CURR:RANG?", ":SOUR:CURR:RANG:AUTO 0;:SOUR:CURR:RANG %g",
-        """ A floating point property that controls the source current
+        """ Control (floating) the source current
         range in Amps, which can take values between -0.105 A and +0.105 A.
         Auto-range is disabled when this property is set. """,
         validator=truncated_range,
@@ -153,7 +153,7 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
     )
     source_auto_range = Instrument.control(
         ":SOUR:CURR:RANG:AUTO?", ":SOUR:CURR:RANG:AUTO %d",
-        """ A boolean property that controls the auto range of the current source.
+        """ Control (boolean) the auto range of the current source.
         Valid values are True or False. """,
         values={True: 1, False: 0},
         map_values=True,
@@ -273,7 +273,7 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     waveform_function = Instrument.control(
         ":SOUR:WAVE:FUNC?", ":SOUR:WAVE:FUNC %s",
-        """ A string property that controls the selected wave function. Valid
+        """ Control (string) the selected wave function. Valid
         values are "sine", "ramp", "square", "arbitrary1", "arbitrary2",
         "arbitrary3" and "arbitrary4". """,
         values={
@@ -290,28 +290,28 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     waveform_frequency = Instrument.control(
         ":SOUR:WAVE:FREQ?", ":SOUR:WAVE:FREQ %g",
-        """A floating point property that controls the frequency of the
+        """Control (floating) the frequency of the
         waveform in Hertz. Valid values are in range 1e-3 to 1e5. """,
         validator=truncated_range,
         values=[1e-3, 1e5]
     )
     waveform_amplitude = Instrument.control(
         ":SOUR:WAVE:AMPL?", ":SOUR:WAVE:AMPL %g",
-        """A floating point property that controls the (peak) amplitude of the
+        """Control (floating) the (peak) amplitude of the
         waveform in Amps. Valid values are in range 2e-12 to 0.105. """,
         validator=truncated_range,
         values=[2e-12, 0.105]
     )
     waveform_offset = Instrument.control(
         ":SOUR:WAVE:OFFS?", ":SOUR:WAVE:OFFS %g",
-        """A floating point property that controls the offset of the waveform
+        """Control (floating) the offset of the waveform
         in Amps. Valid values are in range -0.105 to 0.105. """,
         validator=truncated_range,
         values=[-0.105, 0.105]
     )
     waveform_dutycycle = Instrument.control(
         ":SOUR:WAVE:DCYC?", ":SOUR:WAVE:DCYC %g",
-        """A floating point property that controls the duty-cycle of the
+        """Control (floating) the duty-cycle of the
         waveform in percent for the square and ramp waves. Valid values are in
         range 0 to 100. """,
         validator=truncated_range,
@@ -319,7 +319,7 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
     )
     waveform_duration_time = Instrument.control(
         ":SOUR:WAVE:DUR:TIME?", ":SOUR:WAVE:DUR:TIME %g",
-        """A floating point property that controls the duration of the
+        """Control (floating) the duration of the
         waveform in seconds. Valid values are in range 100e-9 to 999999.999.
         """,
         validator=truncated_range,
@@ -327,7 +327,7 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
     )
     waveform_duration_cycles = Instrument.control(
         ":SOUR:WAVE:DUR:CYCL?", ":SOUR:WAVE:DUR:CYCL %g",
-        """A floating point property that controls the duration of the
+        """Control (floating) the duration of the
         waveform in cycles. Valid values are in range 1e-3 to 99999999900.
         """,
         validator=truncated_range,
@@ -341,14 +341,14 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     waveform_ranging = Instrument.control(
         ":SOUR:WAVE:RANG?", ":SOUR:WAVE:RANG %s",
-        """ A string property that controls the source ranging of the
+        """ Control (string) the source ranging of the
         waveform. Valid values are "best" and "fixed". """,
         values={"best": "BEST", "fixed": "FIX"},
         map_values=True,
     )
     waveform_use_phasemarker = Instrument.control(
         ":SOUR:WAVE:PMAR:STAT?", ":SOUR:WAVE:PMAR:STAT %s",
-        """ A boolean property that controls whether the phase marker option
+        """ Control (boolean) whether the phase marker option
         is turned on or of. Valid values True (on) or False (off). Other
         settings for the phase marker have not yet been implemented.""",
         values={True: 1, False: 0},
@@ -356,13 +356,13 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
     )
     waveform_phasemarker_phase = Instrument.control(
         ":SOUR:WAVE:PMAR?", ":SOUR:WAVE:PMAR %g",
-        """ A numerical property that controls the phase of the phase marker.""",
+        """ Control (numerical) the phase of the phase marker.""",
         validator=truncated_range,
         values=[-180, 180],
     )
     waveform_phasemarker_line = Instrument.control(
         ":SOUR:WAVE:PMAR:OLIN?", ":SOUR:WAVE:PMAR:OLIN %d",
-        """ A numerical property that controls the line of the phase marker.""",
+        """ Control (numerical) the line of the phase marker.""",
         validator=truncated_range,
         values=[1, 6],
     )
@@ -445,7 +445,7 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     display_enabled = Instrument.control(
         ":DISP:ENAB?", ":DISP:ENAB %d",
-        """ A boolean property that controls whether or not the display of the
+        """ Control (boolean) whether or not the display of the
         sourcemeter is enabled. Valid values are True and False. """,
         values={True: 1, False: 0},
         map_values=True,
@@ -453,6 +453,11 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     @property
     def error(self):
+        """Get the next error from the queue.
+
+        .. deprecated:: 0.15
+            Use `next_error` instead.
+        """
         warn("Deprecated to use `error`, use `next_error` instead.", FutureWarning)
         return self.next_error
 
@@ -526,7 +531,7 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     measurement_event_enabled = Instrument.control(
         ":STAT:MEAS:ENAB?", ":STAT:MEAS:ENAB %d",
-        """ An integer value that controls which measurement events are
+        """ Control which measurement events are
         registered in the Measurement Summary Bit (MSB) status bit. Refer to
         the Model 6220/6221 Reference Manual for more information about
         programming the status bits.
@@ -538,7 +543,7 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     operation_event_enabled = Instrument.control(
         ":STAT:OPER:ENAB?", ":STAT:OPER:ENAB %d",
-        """ An integer value that controls which operation events are
+        """ Control which operation events are
         registered in the Operation Summary Bit (OSB) status bit. Refer to
         the Model 6220/6221 Reference Manual for more information about
         programming the status bits.
@@ -550,7 +555,7 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     questionable_event_enabled = Instrument.control(
         ":STAT:QUES:ENAB?", ":STAT:QUES:ENAB %d",
-        """ An integer value that controls which questionable events are
+        """ Control which questionable events are
         registered in the Questionable Summary Bit (QSB) status bit. Refer to
         the Model 6220/6221 Reference Manual for more information about
         programming the status bits.
@@ -562,7 +567,7 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     standard_event_enabled = Instrument.control(
         "ESE?", "ESE %d",
-        """ An integer value that controls which standard events are
+        """ Control which standard events are
         registered in the Event Summary Bit (ESB) status bit. Refer to
         the Model 6220/6221 Reference Manual for more information about
         programming the status bits.
@@ -574,7 +579,7 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     srq_event_enabled = Instrument.control(
         "*SRE?", "*SRE %d",
-        """ An integer value that controls which event registers trigger the
+        """ Control which event registers trigger the
         Service Request (SRQ) status bit. Refer to the Model 6220/6221
         Reference Manual for more information about programming the status
         bits.
@@ -586,7 +591,7 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     measurement_events = Instrument.measurement(
         ":STAT:MEAS?",
-        """ An integer value that reads which measurement events have been
+        """ Get which measurement events have been
         registered in the Measurement event registers. Refer to the Model
         6220/6221 Reference Manual for more information about programming
         the status bits. Reading this value clears the register.
@@ -596,7 +601,7 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     operation_events = Instrument.measurement(
         ":STAT:OPER?",
-        """ An integer value that reads which operation events have been
+        """ Get which operation events have been
         registered in the Operation event registers. Refer to the Model
         6220/6221 Reference Manual for more information about programming
         the status bits. Reading this value clears the register.
@@ -606,7 +611,7 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     questionable_events = Instrument.measurement(
         ":STAT:QUES?",
-        """ An integer value that reads which questionable events have been
+        """ Get which questionable events have been
         registered in the Questionable event registers. Refer to the Model
         6220/6221 Reference Manual for more information about programming
         the status bits. Reading this value clears the register.
@@ -616,7 +621,7 @@ class Keithley6221(KeithleyBuffer, SCPIMixin, Instrument):
 
     standard_events = Instrument.measurement(
         "*ESR?",
-        """ An integer value that reads which standard events have been
+        """ Get which standard events have been
         registered in the Standard event registers. Refer to the Model
         6220/6221 Reference Manual for more information about programming
         the status bits. Reading this value clears the register.